feat(SP3T5): isolar componentes e separar o dash do candidato
